+++ compiler/schema-test.cpp
@@ -189,7 +189,7 @@ TEST(Schema, ReadSimple) {
 
   struct CodeGeneratorRequest req;
   read_CodeGeneratorRequest(&req, root);
-  for (size_t i = 0; i < req.nodes.p.len; i++) {
+  for (int i = 0; i < req.nodes.p.len; i++) {
   }
 }
